Derive Macro enum_index::EnumIndex

#[derive(EnumIndex)]
{
    // Attributes available to this derive:
    #[index_type]
    #[index]
}

A simple derive macro to map static indexs of a single, consistent type to Rust enum variants.

Example

use enum_index::EnumIndex;

#[derive(EnumIndex)]
#[index_type(String)]
pub enum MyEnum {
    #[index("Member 0")]
    MEMBER0,
    #[index("Member 1")]
    MEMBER1,
}

will Expand into:

use enum_index::EnumIndex;
use enum_index_types::errors;

pub enum MyEnum {
    MEMBER0,
    MEMBER1,
}
impl MyEnum {
    fn index(&self) -> String {
        match self {
            Self::MEMBER0 => "Member 0".to_string(),
            Self::MEMBER1 => "Member 1".to_string(),
        }
    }
    fn from_index(index: &str) -> Option<Self> {
        match index {
            "Member 0" => Some(Self::MEMBER0),
            "Member 1" => Some(Self::MEMBER1),
            _ => None,
        }
    }
}
impl From<MyEnum> for String {
    fn from(index: MyEnum) -> Self {
        index.index()
    }
}
impl TryFrom<&str> for MyEnum {
    type Error = errors::EnumIndexError;
    fn try_from(index: &str) -> Result<Self, Self::Error> {
        Self::from_index(&index)
            .ok_or(
                errors::EnumIndexError::IndexNotFound({
                    let res = ::alloc::fmt::format(format_args!("{0:?}", index));
                    res
                }),
            )
    }
}

This allows quick look up from both ways without any use of iteration, while keeping boilerplates to a minimum.

Works with typical primitive types such as u32, usize, i64, char as well as String. &str is supported by String; simply declare the type as String and use str literals for indexs.

Complex types such as Option<usize> will need to be aliased in #[index_type(T)].